requirements.txt
# Dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.50.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Dependencias de desarrollo (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
